+++ dev/setup_venv.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+python setup.py develop
 for package in $(ls -d */); do pushd $package; if [ -e setup.py ]; then python setup.py develop; fi; popd; done
 pip install pip --upgrade
 pip install -r requirements.txt --upgrade