+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+2019-08-23
+==========
+* [Feature] Add bin/ticker.py. It publish a tick event (1 sg, 1 min, 2 min, 5 min and 60 minutres)
+
 2019-07-22
 ==========
 * [Internal] Use serializer from infcommon instead local serializers